+++ AzzyBot-Next/Services/UpdaterService.cs
@@ -1,31 +1,20 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text.Json;
-using System.Threading;
 using System.Threading.Tasks;
-using AzzyBot.Database;
-using AzzyBot.Database.Entities;
 using AzzyBot.Logging;
-using AzzyBot.Services.Interfaces;
-using AzzyBot.Services.Modules;
 using AzzyBot.Settings;
 using AzzyBot.Utilities;
-using AzzyBot.Utilities.Encryption;
 using AzzyBot.Utilities.Records;
-using AzzyBot.Utilities.Records.AzuraCast;
 using DSharpPlus.Entities;
 using Microsoft.Extensions.Logging;
 
 namespace AzzyBot.Services;
 
-public sealed class UpdaterService(ILogger<UpdaterService> logger, IQueuedBackgroundTask taskQueue, AzuraCastApiService azuraCastApiService, AzzyBotSettingsRecord settings, DbActions dbActions, DiscordBotService botService, WebRequestService webService)
+public sealed class UpdaterService(ILogger<UpdaterService> logger, AzzyBotSettingsRecord settings, DiscordBotService botService, WebRequestService webService)
 {
     private readonly ILogger<UpdaterService> _logger = logger;
-    private readonly IQueuedBackgroundTask _taskQueue = taskQueue;
-    private readonly AzuraCastApiService _azuraCastApiService = azuraCastApiService;
     private readonly AzzyBotSettingsRecord _settings = settings;
-    private readonly DbActions _dbActions = dbActions;
     private readonly DiscordBotService _botService = botService;
     private readonly WebRequestService _webService = webService;
     private DateTime _lastAzzyUpdateNotificationTime = DateTime.MinValue;
@@ -33,42 +22,6 @@ public sealed class UpdaterService(ILogger<UpdaterService> logger, IQueuedBackgr
     private int _azzyNotifyCounter;
     private readonly Uri _latestUrl = new("https://api.github.com/repos/Sella-GH/AzzyBot/releases/latest");
     private readonly Uri _previewUrl = new("https://api.github.com/repos/Sella-GH/AzzyBot/releases");
-
-    public async ValueTask QueueAzuraCastUpdatesAsync()
-    {
-        List<GuildsEntity> guilds = await _dbActions.GetGuildsAsync();
-        foreach (AzuraCastEntity azuraCast in guilds.Where(g => g.AzuraCast?.Checks.Updates == true).Select(g => g.AzuraCast!))
-        {
-            _ = Task.Run(async () => await _taskQueue.QueueBackgroundWorkItemAsync(async ct => await CheckForAzuraCastUpdatesAsync(azuraCast, ct)));
-        }
-    }
-
-    private async ValueTask CheckForAzuraCastUpdatesAsync(AzuraCastEntity azuraCast, CancellationToken cancellationToken)
-    {
-        _logger.BackgroundServiceWorkItem();
-
-        cancellationToken.ThrowIfCancellationRequested();
-
-        try
-        {
-            string apiKey = Crypto.Decrypt(azuraCast.AdminApiKey);
-
-            AzuraUpdateRecord update = await _azuraCastApiService.GetUpdatesAsync(new(Crypto.Decrypt(azuraCast.BaseUrl)), apiKey);
-
-            if (!update.NeedsReleaseUpdate && !update.NeedsRollingUpdate)
-                return;
-
-            List<DiscordEmbed> embeds = [EmbedBuilder.BuildAzuraCastUpdatesAvailableEmbed(update)];
-            if (_settings.Updater.DisplayChangelog)
-                embeds.Add(EmbedBuilder.BuildAzuraCastUpdatesChangelogEmbed(update.RollingUpdatesList, update.NeedsRollingUpdate));
-
-            await _botService.SendMessageAsync(azuraCast.NotificationChannelId, null, embeds);
-        }
-        catch (OperationCanceledException)
-        {
-            _logger.OperationCanceled(nameof(CheckForAzuraCastUpdatesAsync));
-        }
-    }
 
     public async Task CheckForAzzyUpdatesAsync()
     {
